Client lourd/requirements.txt
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
ttkbootstrap==1.10.1
pillow==10.1.0
bcrypt==4.0.1
